[main] Attenuate lead channel & improve buffering + final PCB
Lead channel has a high output level, while the signal level after the
Depth control is quite low. Therfore, we need to attenuate the lead
channel before trying to mix in the signal coming from the Depth pot.

Additionnally, the output buffer is changed to a make-up gain stage.

With these changes in mind, the PCB design has been completed.